slight refactoring to the data model
