+++ data/models/dataModel.js
@@ -1,12 +1,19 @@
 const knex = require('knex');
 const knexConfig = require('../knexfile.js');
 const db = knex(knexConfig.development);
- 
-module.exports = {
+
+const db = knex(knexConfig.development);
+ module.exports = {
+    addAction,   
     addProject,
     getProjectActions
 };
- 
+
+function addAction(action) {
+    return db('actions')
+    .insert(action)
+    .into('actions');
+}
 function addProject(project) {
     return db('projects')
     .insert(project)
